Home Screen Redesign
Removed bottom navigation
