+++ app/src/main/java/co/temy/securitysample/HomeActivity.kt
@@ -2,54 +2,57 @@ package co.temy.securitysample
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.BottomNavigationView
-import android.support.design.widget.FloatingActionButton
-import android.support.v4.app.Fragment
-import android.view.MenuItem
-import android.view.View
+import android.util.Log
+import co.temy.securitysample.encryption.CipherWrapper
+import co.temy.securitysample.encryption.KeyStoreWrapper
+import kotlinx.android.synthetic.main.activity_home.*
 
 class HomeActivity : BaseSecureActivity() {
 
-    private lateinit var bottomNavigationView: BottomNavigationView
-    private lateinit var btnAddPassword: FloatingActionButton
+    companion object {
+        val TEST_PASSWORD = "TEST_PASSWORD"
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_home)
 
-        btnAddPassword = findViewById(R.id.btnAddPassword)
-
-        bottomNavigationView = findViewById(R.id.bottomNavigationView)
-        bottomNavigationView.setOnNavigationItemSelectedListener { item -> onNavigationItemSelected(item) }
-
-        showTab(PasswordsFragment())
+        addPasswordView.setOnClickListener { onAddPasswordClick() }
     }
 
-    override fun onStart() {
-        super.onStart()
-    }
-
-    fun onAddPasswordClick(view: View) {
+    private fun onAddPasswordClick() {
         val intent = Intent(this, AddSecretActivity::class.java)
         startActivity(intent)
     }
 
-    private fun onNavigationItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == R.id.menuPasswords) {
-            showTab(PasswordsFragment())
-            btnAddPassword.show()
-            return true
-        } else if (item.itemId == R.id.menuEncryption) {
-            showTab(EncryptionFragment())
-            btnAddPassword.hide()
-            return true
-        }
-        return false
+    private fun printAliases() {
+        val keyStore = KeyStoreWrapper()
+        keyStore.getAllKeyAliases()
+                .sortedBy { it.creationDate }
+                .forEach { Log.i("EncryptionFragment", "${it} \n") }
     }
 
-    private fun showTab(tab: Fragment) {
-        val fragmentTransaction = supportFragmentManager.beginTransaction()
-        fragmentTransaction.replace(R.id.tabContainer, tab)
-        fragmentTransaction.commitNow()
+    private var keyNumber = 0
+
+    private fun createKey() {
+        val keyStore = KeyStoreWrapper()
+        keyStore.createSymmetricKey("Test-${++keyNumber}", false, false)
+    }
+
+    private fun testEncryption() {
+        val keyStore = KeyStoreWrapper()
+        val symmetricKey = keyStore.createSymmetricKey(alias = TEST_PASSWORD, userAuthenticationRequired = false, invalidatedByBiometricEnrollment = false)
+
+        val cipher = CipherWrapper(CipherWrapper.TRANSFORMATION_SYMMETRIC)
+
+        if (symmetricKey != null) {
+
+            val encryptedData = cipher.encrypt(TEST_PASSWORD, symmetricKey)
+            val decryptedData = cipher.decrypt(encryptedData, symmetricKey)
+
+            Log.i("EncryptionFragment", "Input Value : ${TEST_PASSWORD}")
+            Log.i("EncryptionFragment", "Encrypted Value : $encryptedData")
+            Log.i("EncryptionFragment", "Decrypted Value : ${decryptedData}")
+        }
     }
 }
