--- app/src/main/java/co/temy/securitysample/PasswordsFragment.kt
@@ -1,13 +1,0 @@
-package co.temy.securitysample;
-
-import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class PasswordsFragment : Fragment() {
-    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater?.inflate(R.layout.fragment_passwords, container, false)
-    }
-}
