+++ app/src/main/res/layout/activity_home.xml
@@ -8,40 +8,21 @@
         android:id="@+id/rootView"
         tools:context="co.temy.securitysample.HomeActivity">
 
-    <FrameLayout
-            android:layout_width="0dp"
-            android:layout_height="0dp" android:layout_marginTop="0dp"
-            app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginStart="0dp" app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="0dp"
-            android:layout_marginBottom="0dp" app:layout_constraintBottom_toTopOf="@+id/bottomNavigationView"
-            android:id="@+id/tabContainer">
-    </FrameLayout>
-
-    <android.support.design.widget.BottomNavigationView
-            android:id="@+id/bottomNavigationView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:menu="@menu/bottom_navigation"
-            app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintStart_toStartOf="parent"/>
 
     <android.support.design.widget.FloatingActionButton
             app:useCompatPadding="true"
             android:src="@drawable/ic_add_white"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:clickable="true"
-            android:id="@+id/btnAddPassword" app:fabSize="normal" app:layout_constraintEnd_toEndOf="parent"
-            android:layout_marginEnd="0dp"
-            android:onClick="onAddPasswordClick"
-            android:layout_marginBottom="0dp" app:layout_constraintBottom_toTopOf="@+id/bottomNavigationView"/>
-    <View
-            android:layout_width="0dp"
-            android:background="@drawable/shadow_bg"
-            android:layout_height="1dp"
-            android:id="@+id/shadow"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0" android:layout_marginBottom="0dp"
-            app:layout_constraintBottom_toTopOf="@+id/bottomNavigationView"/>
-
+            android:id="@+id/addPasswordView" app:fabSize="normal" app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginBottom="0dp" app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_marginEnd="0dp"/>
+    <TextView
+            android:text="@string/home_empty_view_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/emptyView" android:layout_marginTop="8dp"
+            app:layout_constraintTop_toTopOf="parent" app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_marginBottom="8dp" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="8dp" app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"/>
 </android.support.constraint.ConstraintLayout>
