--- app/src/main/res/layout/fragment_encryption.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context="co.temy.securitysample.HomeActivity">
-
-    <Button
-            android:text="ALIASES"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/btnAliases" app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintTop_toTopOf="parent" android:layout_marginTop="8dp"
-            style="@style/Widget.AppCompat.Button.Colored"/>
-    <Button
-            android:text="CREATE KEY"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/btnCreateSyncKey" app:layout_constraintStart_toEndOf="@+id/btnAliases"
-            android:layout_marginStart="8dp" app:layout_constraintTop_toTopOf="parent" android:layout_marginTop="8dp"
-            style="@style/Widget.AppCompat.Button.Colored"/>
-    <Button
-            android:text="ENCRYPT"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/btnEncrypt" app:layout_constraintStart_toEndOf="@+id/btnCreateSyncKey"
-            android:layout_marginStart="8dp" app:layout_constraintTop_toTopOf="parent" android:layout_marginTop="8dp"
-            style="@style/Widget.AppCompat.Button.Colored"/>
-
-</android.support.constraint.ConstraintLayout>
