--- app/src/main/res/menu/bottom_navigation.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-            android:id="@+id/menuPasswords"
-            android:title="@string/home_passwords_tab_title"
-            android:icon="@drawable/ic_home_white"/>
-    <item
-            android:id="@+id/menuEncryption"
-            android:title="@string/home_encryption_tab_title"
-            android:icon="@drawable/ic_fingerprint_white"/>
-</menu>
